Update presentation for cleaner style
The look was revamped substantially.

After using Markdown Preview Enhanced within Visual Studio Code, the
look was simply too appealing to resist using in this project. To pick
and choose what was desired, we used inspect to mark all the styling per
property to emulate it as much as possible.

In doing so, syntax highlighting is now broken as Jekyll is no longer
being relied on for this to maintain styling consistently and appropria-
tely. Prism.js is the planned replacement at this time.

Attribution has been included and thanks go out to the contributors of
Markdown Preview Enhanced for making such a gorgeous style.