Improve consistency
The previous commit (a3bd6ed) aimed to improve consistency. However, it
didn't go far enough: it used a path for a given blog post that was set
in a manner that conflicted with the oft-seen method of utilizing the
hyphenated title of the post.

This has now been remedied.